Drop Language.isWiderComm
I believe the thought here was that "wider communication" was a standardized piece of info that was declared, agreed upon, externally.
And then this flag would limit the languages allowed for `Partner.languageOfWiderCommunication`.

After reviewing with Seth/business, this relationship could really be to any language we have. And because of this there's no need to restrict.
And now I'm doubting that this flag would serve the purpose, in practice, that I described above.