Move edit departmentId permission check to policies (#3547)
This corrects to covey `canEdit` correctly.

This does remove the impersonation introspection (check if the current user is actually an admin impersonating someone else), but that hasn't been needed, and is arguably a bad practice.

FYI the check in create is still the same as this was before. We do actually use that in the monday sync.
We don't actually check edit permission for any properties on create with the permission system. I think it is fine to stay how it is, because we don't expose UI for that.

Related to: #3383