Better handle validation errors
There's still a weird case where the first time
a validation error is thrown, it does not have the expected extensions on it.
But the next time it does.
This is observed in the `onValidate` plugin step.
There's gotta be a cache that is re-used, and maybe some internal plugins that are happening after ours to add code/http to extension.