Fix subscription field resolvers loosing async context
This happened because the subscription resolver returns an iterator.
This iteration (aka waiting for and processing events) happens
after/outside the subscribe() function.

We fix by wrapping the iterator so that this iteration re-enters
into the async context when grabbing the next item/event that
the iterator was created with.