Move subscription observable normalization up in call chain
Same logic as before, but in one layer up.
This allows us to call the subscription resolver methods ourselves.
e.g., to merge multiple together.

Previously we were changing the result of the direct resolver methods,
modifying the signature. This was done, magically, unknown to TS.
This meant that if we called a resolver method ourselves, we would have
a broken stream & expectation: We'd consume as an Observable,
but it would actually be an AsyncIterable.

This new way we don't have to modify the signature of the resolver.
Instead, we apply this conversion on the layer that calls the resolver.

Unfortunately, this is a monkey patch, but at least it's not yarn patch.